refactor(priceAuthority): move to api/priceAuthority
