+++ api/priceAuthority/invert.js
@@ -1,3 +1,4 @@
+/* global process */
 // @ts-check
 import { E } from '@agoric/far';
 import '@agoric/zoe/exported';
@@ -7,9 +8,6 @@ import '@agoric/zoe/src/contracts/exported';
  * @typedef {Object} DeployPowers The special powers that `agoric deploy` gives us
  * @property {(path: string) => Promise<{ moduleFormat: string, source: string }>} bundleSource
  * @property {(path: string) => string} pathResolve
- * @property {(path: string, opts?: any) => Promise<any>} installUnsafePlugin
- * @property {string} host
- * @property {string} port
  *
  * @typedef {Object} Board
  * @property {(id: string) => any} getValue
@@ -21,9 +19,12 @@ import '@agoric/zoe/src/contracts/exported';
 /**
  * @typedef {{ board: Board, chainTimerService, wallet, scratch, spawner }} Home
  * @param {Promise<Home>} homePromise
+ * @param {Object} root0
+ * @param {(filename: string) => Promise<any>} root0.bundleSource
+ * @param {(filename: string) => string} root0.pathResolve
  * A promise for the references available from REPL home
  */
-export default async function priceAuthorityfromNotifier(
+export default async function priceAuthorityInvert(
   homePromise,
   { bundleSource, pathResolve },
 ) {
@@ -31,11 +32,10 @@ export default async function priceAuthorityfromNotifier(
     FORCE_SPAWN = 'true',
     IN_ISSUER_JSON = JSON.stringify('Testnet.$LINK'),
     OUT_ISSUER_JSON = JSON.stringify('Testnet.$USD'),
-    PRICE_DECIMALS = '0',
-    NOTIFIER_BOARD_ID,
+    PRICE_AUTHORITY_BOARD_ID,
   } = process.env;
 
-  if (!NOTIFIER_BOARD_ID) {
+  if (!PRICE_AUTHORITY_BOARD_ID) {
     console.error(`You must specify PRICE_AUTHORITY_BOARD_ID`);
     process.exit(1);
   }
@@ -71,26 +71,11 @@ export default async function priceAuthorityfromNotifier(
     process.exit(1);
   }
 
-  const displayInfoIn = await E(E(issuerIn).getBrand()).getDisplayInfo();
-  const { decimalPlaces: decimalPlacesIn = 0 } = displayInfoIn || {};
-
-  const unitValueIn = 10 ** decimalPlacesIn;
-
-  const displayInfoOut = await E(E(issuerOut).getBrand()).getDisplayInfo();
-  const { decimalPlaces: decimalPlacesOut = 0 } = displayInfoOut || {};
-
-  // Take a price with priceDecimalPlaces and scale it to have decimalPlacesOut.
-  const priceDecimalPlaces = JSON.parse(PRICE_DECIMALS);
-  const scaleValueOut = 10 ** (decimalPlacesOut - priceDecimalPlaces);
-
-  // Get the notifier.
-  const notifierId = NOTIFIER_BOARD_ID.replace(/^board:/, '');
-  const notifier = E(board).getValue(notifierId);
-
+  const priceAuthority = await E(board).getValue(PRICE_AUTHORITY_BOARD_ID);
   let priceAuthorityFactory = E(scratch).get('priceAuthorityFactory');
 
   if (FORCE_SPAWN || !priceAuthorityFactory) {
-    // Bundle up the notifierPriceAuthority code
+    // Bundle up the priceAuthorityFactory code
     const bundle = await bundleSource(
       pathResolve('./src/priceAuthorityFactory.js'),
     );
@@ -106,17 +91,20 @@ export default async function priceAuthorityfromNotifier(
   }
 
   console.log('Waiting for first valid quote from push notifier...');
-  const priceAuthority = await E(
+  const inversePriceAuthority = await E(
     priceAuthorityFactory,
-  ).makeNotifierPriceAuthority({
-    notifier,
+  ).makeInversePriceAuthority({
+    priceAuthority,
     issuerIn,
     issuerOut,
     timer,
-    unitValueIn,
-    scaleValueOut,
   });
 
-  const PRICE_AUTHORITY_BOARD_ID = await E(board).getId(priceAuthority);
-  console.log('-- PRICE_AUTHORITY_BOARD_ID:', PRICE_AUTHORITY_BOARD_ID);
+  const INVERSE_PRICE_AUTHORITY_BOARD_ID = await E(board).getId(
+    inversePriceAuthority,
+  );
+  console.log(
+    '-- INVERSE_PRICE_AUTHORITY_BOARD_ID:',
+    INVERSE_PRICE_AUTHORITY_BOARD_ID,
+  );
 }
