+++ api/priceAuthority/src/registry.js
@@ -1,0 +1,1 @@
+export { makePriceAuthorityRegistry } from '@agoric/zoe/tools/priceAuthorityRegistry.js';
