--- priceAuthority/src/index.js
@@ -1,5 +1,0 @@
-export * from './fake';
-export * from './linear';
-export * from './inverseQuotes';
-export * from './registry';
-export * from './single';
